Extract arbitrary system generation into test_utils
