+++ .idea/recs.iml
@@ -12,6 +12,7 @@
       <sourceFolder url="file://$MODULE_DIR$/crates/ecs/tests" isTestSource="true" />
       <sourceFolder url="file://$MODULE_DIR$/crates/scheduler/src" isTestSource="false" />
       <sourceFolder url="file://$MODULE_DIR$/crates/scheduler/tests" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/crates/test_utils/src" isTestSource="false" />
       <excludeFolder url="file://$MODULE_DIR$/target" />
       <excludeFolder url="file://$MODULE_DIR$/crates/ecs/benches/target" />
     </content>
