+++ crates/ecs/src/lib.rs
@@ -674,10 +674,8 @@ impl_system_parameter_function!(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14
 #[cfg(test)]
 mod tests {
     use super::*;
-    use proptest::prop_compose;
     use test_case::test_case;
     use test_log::test;
-    use test_strategy::proptest;
 
     #[derive(Debug)]
     struct A;
@@ -752,80 +750,5 @@ mod tests {
         let component_accesses = system.into_system().component_accesses();
 
         assert_eq!(expected_accesses, component_accesses)
-    }
-
-    // This might be useful for a wide variety of tests in the project, maybe
-    // worth extracting to test_system crate along with all of the prop_compose calls?
-    #[derive(Debug)]
-    struct MockSystem {
-        name: String,
-        parameters: Vec<ComponentAccessDescriptor>,
-    }
-
-    impl System for MockSystem {
-        fn name(&self) -> &str {
-            &self.name
-        }
-
-        fn run(&self, _world: &World) {
-            panic!("mocked system, not meant to be run")
-        }
-
-        fn component_accesses(&self) -> Vec<ComponentAccessDescriptor> {
-            self.parameters.clone()
-        }
-    }
-
-    prop_compose! {
-        fn arb_component_type()
-                             (type_index in 0_usize..8)
-                             -> TypeId {
-            let types = vec![
-                TypeId::of::<u8>(),
-                TypeId::of::<i8>(),
-                TypeId::of::<u16>(),
-                TypeId::of::<i16>(),
-                TypeId::of::<u32>(),
-                TypeId::of::<i32>(),
-                TypeId::of::<u64>(),
-                TypeId::of::<i64>(),
-            ];
-            types[type_index]
-        }
-    }
-
-    prop_compose! {
-        fn arb_component_access()
-                               (write in proptest::arbitrary::any::<bool>(),
-                                type_id in arb_component_type())
-                               -> ComponentAccessDescriptor {
-            if write {
-                ComponentAccessDescriptor::Write(type_id)
-            } else {
-                ComponentAccessDescriptor::Read(type_id)
-            }
-        }
-    }
-
-    prop_compose! {
-        #[allow(trivial_casts)] // Compiler won't coerce `MockSystem` to `dyn System` for some reason.
-        fn arb_system()
-                     (name in proptest::arbitrary::any::<String>(),
-                      parameters in proptest::collection::vec(arb_component_access(), 1..8))
-                     -> Box<dyn System> {
-            Box::new(MockSystem {
-                name,
-                parameters,
-            }) as Box<dyn System>
-        }
-    }
-
-    #[proptest]
-    fn systems_fulfill_equivalence_relation(
-        #[strategy(arb_system())] a: Box<dyn System>,
-        #[strategy(arb_system())] b: Box<dyn System>,
-        #[strategy(arb_system())] c: Box<dyn System>,
-    ) {
-        reltester::eq(&a, &b, &c).unwrap()
     }
 }
