+++ crates/scheduler/Cargo.toml
@@ -18,6 +18,7 @@ itertools = { workspace = true }        # Iterator helper functions
 daggy = { workspace = true }            # Directed Acyclic Graphs (DAGs)
 
 [dev-dependencies]
+test_utils = { path = "../test_utils" }
 criterion = { workspace = true }        # For benchmarking
 proptest = { workspace = true }         # Property-based testing tools and harness
 test-strategy = { workspace = true }    # Prop-test macros that are more ergonomic than the ones from proptest
