+++ crates/test_utils/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "test_utils"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ecs = { path = "../ecs" }
+proptest = { workspace = true }         # Property-based testing tools and harness
